update atomix.js and add best solutions
The x library functions are now in the atomix.js file
instead of a separate file and are included in the
KP_ATOMIX namespace instead of the global namespace.


Added the best solutions as javascript files (actually,
JSON with a javascript header).  These are used by
the html best solutions pages, together with solutions.js
to build the html tables.